refactor: 회원 가입 기능을 Auth 에서 Member 기능으로 이동
