+++ src/main/java/woowacourse/auth/application/AuthService.java
@@ -3,16 +3,15 @@ package woowacourse.auth.application;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import woowacourse.auth.dao.MemberDao;
-import woowacourse.auth.domain.Email;
 import woowacourse.auth.domain.Member;
 import woowacourse.auth.dto.request.LoginRequest;
-import woowacourse.auth.dto.request.MemberCreateRequest;
 import woowacourse.auth.dto.request.PasswordCheckRequest;
 import woowacourse.auth.dto.response.LoginResponse;
 import woowacourse.auth.dto.response.PasswordCheckResponse;
 import woowacourse.auth.support.JwtTokenProvider;
 
 @Service
+@Transactional(readOnly = true)
 public class AuthService {
 
     private final MemberDao memberDao;
@@ -23,27 +22,6 @@ public class AuthService {
         this.jwtTokenProvider = jwtTokenProvider;
     }
 
-    @Transactional
-    public Long save(MemberCreateRequest memberCreateRequest) {
-        validateUniqueEmail(memberCreateRequest);
-        Member member = new Member(memberCreateRequest.getEmail(), memberCreateRequest.getPassword(),
-                memberCreateRequest.getNickname());
-        return memberDao.save(member);
-    }
-
-    private void validateUniqueEmail(MemberCreateRequest memberCreateRequest) {
-        if (existsEmail(memberCreateRequest.getEmail())) {
-            throw new IllegalArgumentException("이미 존재하는 이메일 주소입니다.");
-        }
-    }
-
-    @Transactional(readOnly = true)
-    public boolean existsEmail(String email) {
-        String validatedEmail = new Email(email).getValue();
-        return memberDao.existsEmail(validatedEmail);
-    }
-
-    @Transactional(readOnly = true)
     public LoginResponse login(LoginRequest loginRequest) {
         Member member = memberDao.findByEmail(loginRequest.getEmail())
                 .orElseThrow(() -> new IllegalArgumentException("이메일과 비밀번호를 확인해주세요."));
@@ -58,7 +36,6 @@ public class AuthService {
         }
     }
 
-    @Transactional(readOnly = true)
     public PasswordCheckResponse checkPassword(Long memberId, PasswordCheckRequest passwordCheckRequest) {
         Member member = memberDao.findById(memberId)
                 .orElseThrow(() -> new IllegalArgumentException("이메일과 비밀번호를 확인해주세요."));
