+++ src/main/java/woowacourse/auth/application/MemberService.java
@@ -7,6 +7,7 @@ import woowacourse.auth.domain.Email;
 import woowacourse.auth.domain.Member;
 import woowacourse.auth.domain.Nickname;
 import woowacourse.auth.domain.Password;
+import woowacourse.auth.dto.request.MemberCreateRequest;
 import woowacourse.auth.dto.request.MemberUpdateRequest;
 import woowacourse.auth.dto.request.PasswordUpdateRequest;
 import woowacourse.auth.dto.response.MemberResponse;
@@ -22,16 +23,24 @@ public class MemberService {
         this.memberDao = memberDao;
     }
 
+    @Transactional
+    public Long save(MemberCreateRequest memberCreateRequest) {
+        validateUniqueEmail(memberCreateRequest);
+        Member member = new Member(memberCreateRequest.getEmail(), memberCreateRequest.getPassword(),
+                memberCreateRequest.getNickname());
+        return memberDao.save(member);
+    }
+
+    private void validateUniqueEmail(MemberCreateRequest memberCreateRequest) {
+        if (existsEmail(memberCreateRequest.getEmail())) {
+            throw new IllegalArgumentException("이미 존재하는 이메일 주소입니다.");
+        }
+    }
+
     public boolean existsEmail(String email) {
         String validatedEmail = new Email(email).getValue();
         return memberDao.existsEmail(validatedEmail);
     }
-
-//    public MemberResponse find(String email) {
-//        Member member = memberDao.findByEmail(email)
-//                .orElseThrow(() -> new AuthorizationException("회원 정보를 찾지 못했습니다."));
-//        return new MemberResponse(member);
-//    }
 
     public MemberResponse find(Long id) {
         Member member = memberDao.findById(id)
