+++ src/main/java/woowacourse/auth/ui/AuthController.java
@@ -1,7 +1,6 @@
 package woowacourse.auth.ui;
 
 import javax.validation.Valid;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -11,7 +10,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import woowacourse.auth.application.AuthService;
 import woowacourse.auth.dto.request.LoginRequest;
-import woowacourse.auth.dto.request.MemberCreateRequest;
 import woowacourse.auth.dto.request.PasswordCheckRequest;
 import woowacourse.auth.dto.response.EmailUniqueCheckResponse;
 import woowacourse.auth.dto.response.LoginResponse;
@@ -26,19 +24,6 @@ public class AuthController {
 
     public AuthController(AuthService authService) {
         this.authService = authService;
-    }
-
-    @PostMapping("/members")
-    public ResponseEntity<Void> signUp(@RequestBody @Valid MemberCreateRequest memberCreateRequest) {
-        authService.save(memberCreateRequest);
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .build();
-    }
-
-    @GetMapping("/members/email-check")
-    public ResponseEntity<EmailUniqueCheckResponse> checkUniqueEmail(@RequestParam String email) {
-        EmailUniqueCheckResponse emailUniqueCheckResponse = new EmailUniqueCheckResponse(!authService.existsEmail(email));
-        return ResponseEntity.ok(emailUniqueCheckResponse);
     }
 
     @PostMapping("/login")
