+++ src/main/java/woowacourse/auth/ui/MemberController.java
@@ -1,16 +1,21 @@
 package woowacourse.auth.ui;
 
 import javax.validation.Valid;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import woowacourse.auth.application.MemberService;
+import woowacourse.auth.dto.request.MemberCreateRequest;
 import woowacourse.auth.dto.request.MemberUpdateRequest;
 import woowacourse.auth.dto.request.PasswordUpdateRequest;
+import woowacourse.auth.dto.response.EmailUniqueCheckResponse;
 import woowacourse.auth.dto.response.MemberResponse;
 import woowacourse.auth.support.AuthenticationPrincipal;
 
@@ -22,6 +27,19 @@ public class MemberController {
 
     public MemberController(MemberService memberService) {
         this.memberService = memberService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Void> signUp(@RequestBody @Valid MemberCreateRequest memberCreateRequest) {
+        memberService.save(memberCreateRequest);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .build();
+    }
+
+    @GetMapping("/email-check")
+    public ResponseEntity<EmailUniqueCheckResponse> checkUniqueEmail(@RequestParam String email) {
+        EmailUniqueCheckResponse emailUniqueCheckResponse = new EmailUniqueCheckResponse(!memberService.existsEmail(email));
+        return ResponseEntity.ok(emailUniqueCheckResponse);
     }
 
     @GetMapping("/me")
