+++ src/test/java/woowacourse/auth/application/AuthServiceTest.java
@@ -21,32 +21,14 @@ class AuthServiceTest {
 
     @Autowired
     private AuthService authService;
-
-    @DisplayName("회원 객체를 생성하고 DB에 저장한다.")
-    @Test
-    void saveMember() {
-        MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
-
-        authService.save(memberCreateRequest);
-    }
-
-    @DisplayName("이미 존재하는 이메일로 회원을 생성하려고 하면 예외를 반환한다.")
-    @Test
-    void saveMember_DuplicatedEmail() {
-        MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
-
-        authService.save(memberCreateRequest);
-
-        assertThatThrownBy(() -> authService.save(memberCreateRequest))
-                .isInstanceOf(IllegalArgumentException.class)
-                .hasMessage("이미 존재하는 이메일 주소입니다.");
-    }
+    @Autowired
+    private MemberService memberService;
 
     @DisplayName("로그인에 성공하면 토큰과 닉네임을 반환한다.")
     @Test
     void login() {
         MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
-        authService.save(memberCreateRequest);
+        memberService.save(memberCreateRequest);
         LoginRequest loginRequest = new LoginRequest("abc@woowahan.com", "1q2w3e4r!");
 
         LoginResponse loginResponse = authService.login(loginRequest);
@@ -60,7 +42,7 @@ class AuthServiceTest {
     @CsvSource({"abc@naver.com, 1q2w3e4r!", "abc@woowahan.com, asdas1123!", "abc@naver.com, asdas1123!"})
     void login_Invalid(String email, String password) {
         MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
-        authService.save(memberCreateRequest);
+        memberService.save(memberCreateRequest);
         LoginRequest loginRequest = new LoginRequest(email, password);
 
         assertThatThrownBy(() -> authService.login(loginRequest))
@@ -73,7 +55,7 @@ class AuthServiceTest {
     @CsvSource({"1q2w3e4r!, true", "asda1234!, false"})
     void checkPassword(String password, boolean expected) {
         MemberCreateRequest memberCreateRequest = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "닉네임");
-        Long memberId = authService.save(memberCreateRequest);
+        Long memberId = memberService.save(memberCreateRequest);
         PasswordCheckRequest passwordCheckRequest = new PasswordCheckRequest(password);
 
         boolean actual = authService.checkPassword(memberId, passwordCheckRequest)
