+++ src/test/java/woowacourse/auth/ui/MemberControllerTest.java
@@ -9,11 +9,14 @@ import static org.mockito.Mockito.verify;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -22,6 +25,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.HandlerInterceptor;
 import woowacourse.auth.application.MemberService;
+import woowacourse.auth.dto.request.MemberCreateRequest;
 import woowacourse.auth.dto.request.MemberUpdateRequest;
 import woowacourse.auth.dto.request.PasswordUpdateRequest;
 import woowacourse.auth.dto.response.MemberResponse;
@@ -47,6 +51,52 @@ public class MemberControllerTest {
     private MockMvc mockMvc;
     @Autowired
     private ObjectMapper objectMapper;
+
+    @DisplayName("회원 가입에 성공한다. - 201 Created")
+    @Test
+    void signUp_Created() throws Exception {
+        MemberCreateRequest requestBody = new MemberCreateRequest("abc@woowahan.com", "1q2w3e4r!", "우아한");
+
+        given(memberService.save(any()))
+                .willReturn(MEMBER_ID);
+
+        mockMvc.perform(post("/api/members")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(requestBody)))
+                .andExpect(status().isCreated());
+
+        verify(memberService, times(1))
+                .save(any());
+    }
+
+    @DisplayName("형식에 맞지 않는 정보로 회원 가입에 실패한다. - 400 Bad Request")
+    @ParameterizedTest
+    @CsvSource({"abc,1q2w3e4r!,닉네임", "abc@woowahan.com,1q2w3e4r,닉네임", "abc@woowahan.com,1q2w3e4r!,잘못된닉네임"})
+    void signUp_BadRequest(String email, String password, String nickname) throws Exception {
+        MemberCreateRequest requestBody = new MemberCreateRequest(email, password, nickname);
+
+        given(memberService.save(any()))
+                .willReturn(MEMBER_ID);
+
+        mockMvc.perform(post("/api/members")
+                        .contentType(MediaType.APPLICATION_JSON)
+                        .content(objectMapper.writeValueAsString(requestBody)))
+                .andExpect(status().isBadRequest());
+        verify(memberService, times(0))
+                .save(any());
+    }
+
+    @DisplayName("이메일 중복 체크를 한다. - 200 Ok")
+    @Test
+    void checkUniqueEmail() throws Exception {
+        given(memberService.existsEmail(any()))
+                .willReturn(true);
+
+        mockMvc.perform(get("/api/members/email-check?email=" + "abc@woowahan.com"))
+                .andExpect(status().isOk());
+        verify(memberService, times(1))
+                .existsEmail("abc@woowahan.com");
+    }
 
     @DisplayName("회원 정보를 조회한다. - 200 Ok")
     @Test
